requirements.txt
# Free Consent Management System - Dependencies

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# PDF Processing & OCR
pytesseract==0.3.10
pdf2image==1.16.3
pdfplumber==0.10.3
PyPDF2==3.0.1
Pillow==10.1.0

# AI/LLM (Local)
ollama==0.1.6

# Database
SQLAlchemy==2.0.23

# Security
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
